requirements.txt
asgiref==3.4.1
Django==4.0.1
mutagen==1.45.1
Pillow==9.0.0
psycopg2==2.9.3
pycparser==2.21
python-magic-bin==0.4.14
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
